requirements.txt
anyio==3.6.2
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
